% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleed1yr.R
\docType{data}
\name{bleed1yr}
\alias{bleed1yr}
\title{One-year risk of bleeding}
\format{A data frame with 2560 rows and 16 variables:
\describe{
  \item{year2000}{xxx}
  \item{age}{Age of patient}
  \item{bleeding}{xxx}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}
  \item{}{}

  ...
}}
\source{
\url{https://hjerteforeningen.dk/alt-om-dit-hjerte/}
}
\usage{
bleed1yr
}
\description{
A dataset containing the 1-year risk of bleeding for all combinations of the
risk factors in untreated patients.
}
\keyword{datasets}
