% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stroke1yr.R
\docType{data}
\name{stroke1yr}
\alias{stroke1yr}
\title{One-year risk of stroke}
\format{A data frame with 5120 rows and 16 variables:
\describe{
  \item{year2000}{xxx}
  \item{age}{Age of patient}
  \item{heartfailure}{xxx}
  \item{hypertension}{}
  \item{diabetes}{}
  \item{vascular}{}
  \item{female}{}
  \item{stroke}{}
  \item{chadsvasc}{}
  \item{stroke1y}{}
  \item{stroke1y.lower}{}
  \item{stroke1y.upper}{}
  \item{I63}{}
  \item{I63.lower}{}
  \item{I63.upper}{}
  \item{thromb1y}{}
  \item{thromb1y.lower}{}
  \item{thromb1y.upper}{}

  ...
}}
\source{
\url{https://hjerteforeningen.dk/alt-om-dit-hjerte/}
}
\usage{
stroke1yr
}
\description{
A dataset containing the 1-year risk of stroke for all combinations of the
risk factors in untreated patients.
}
\keyword{datasets}
