% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bleed1yr.R
\docType{data}
\name{bleed1yr}
\alias{bleed1yr}
\title{One-year risk of bleeding}
\format{A data frame with 2560 rows and 16 variables:
\describe{
  \item{year2000}{Year the prediction is valid for. 15 means 2015.}
  \item{age}{Age of patient}
  \item{bleeding}{Previous history of bleeding}
  \item{stroke}{}
  \item{hypertension}{}
  \item{ckd}{}
  \item{drugs}{}
  \item{bleeding1y}{}
  \item{bleeding1y.lower}{}
  \item{bleeding1y.upper}{}
  \item{ICH1y}{}
  \item{ICH1y.lower}{}
  \item{ICH1y.upper}{}
  \item{GI1y}{}
  \item{GI1y.lower}{}
  \item{GI1y.upper}{}

  ...
}}
\source{
\url{https://hjerteforeningen.dk/alt-om-dit-hjerte/}
}
\usage{
bleed1yr
}
\description{
A dataset containing the 1-year risk of bleeding for all combinations of the
risk factors in untreated patients.
}
\keyword{datasets}
